+++ sass/components/_sortButton.scss
@@ -1,4 +1,4 @@
 .sort {
   display: flex;
-  text-align: left;
+  justify-content: flex-start;
 }
