+++ sass/layout/_likeCounter.scss
@@ -1,0 +1,8 @@
+.likes__counter {
+  position: sticky;
+  bottom: 0rem;
+  left: 70rem;
+  background-color: #D3573C;
+  padding: 2rem;
+
+}
