+++ sass/main.scss
@@ -12,6 +12,7 @@
 // Layout
 @import "./layout/photographerInfo";
 @import "./layout/media";
+@import "./layout/likeCounter";
 
 // Components
 @import "./components/button";
